libs/admin/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minxinagent-admin"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "aiosqlite>=0.19.0,<1.0.0",
    "langgraph-checkpoint-sqlite>=2.0.0,<4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["admin_app"]
